avoid wrap-around when scheduling timer ops
git-svn-id: http://svn.vuze.com/public/client@42810 c2c1b7af-1477-405b-86ee-8aefd1fc4dd2

+++ trunk/azureus2/src/com/aelitis/azureus/plugins/magnet/MagnetPlugin.java
@@ -983,7 +983,7 @@ MagnetPlugin
 			
 			if ( secs <= 0 ){
 				
-				timeout = Long.MAX_VALUE;
+				timeout = Integer.MAX_VALUE;
 				
 			}else{
